github/pyproject.toml
[tool.poetry]
name = "github"
version = "0.1.0"
description = ""
authors = ["Scott Mountenay <scott@lightsonsoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = {version = "2.2.5", extras = ["swagger-ui"]}
requests = "^2.31.0"
connexion = {version = "2.14.2", extras = ["swagger-ui"]}
python-dotenv = "^1.0.0"
mypy = "^1.4.1"
gunicorn = "^21.2.0"


[tool.poetry.group.development.dependencies]
black = "^23.7.0"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
